remove needless casts for performance

+++ src/dgtk.py
@@ -42,13 +42,13 @@ import gtk
 ## Cell data functions to pass to add_func_column()
 
 def cell_data_speed(column, cell, model, iter, data):
-    speed = int(model.get_value(iter, data))
+    speed = model.get_value(iter, data)
     if speed<100: speed_str = ""
     else: speed_str = common.fspeed(speed)
     cell.set_property('text', speed_str)
 
 def cell_data_size(column, cell, model, iter, data):
-	size = long(model.get_value(iter, data))
+	size = model.get_value(iter, data)
 	size_str = common.fsize(size)
 	cell.set_property('text', size_str)
